Add New .closure case to SFOnboardingControl.Action
Currently, when a user intends to have a custom closure called when a particular action is performed a '.void' or '.callback' Action case must be used. From there, the associated Action case's selector is used to create a one-to-one mapping with the CommunicationClosure provided. However, this creates a significant issue when attempting to have multiple closures called for the same parent Action case. By adding the '.closure' case to Action, the issue will be fixed properly and create an opportunity for the depreciation and removal of the '.void' and '.callback' Action cases.

In deprecating the '.callback' and '.void' Action cases, the 'callbackTable' property is also deprecated and should be removed at a later point in time.

In order to implement the new '.closure' Action case, the Objective-C runtime is used for the dynamic manipulation of implementation and class method call stacks. A closure block is created for use with objc, and it invokes the closure initially provided, while utilizing weak self in order to prevent reference loops. The block is used for creating a objc implementation that is added to the class for a selector with a UUID string.

Callback execution is no longer supported and thus has been removed from the implementation. The 'void(_:)' method has also been removed.